Update dependencies
* Fix tests
* Modify functionality to not try to render undefined functions
* Fix cyclical dependency

closes #24
closes #25
