+++ package.json
@@ -27,8 +27,8 @@
     "clean": "backstage-cli clean"
   },
   "dependencies": {
-    "@backstage/catalog-model": "^0.7.1",
-    "@backstage/core": "^0.6.0",
+    "@backstage/catalog-model": "^0.7.4",
+    "@backstage/core": "^0.7.3",
     "@material-ui/core": "^4.9.10",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.49",
@@ -41,9 +41,9 @@
     "react-use": "^15.3.6"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.4.0",
-    "@backstage/dev-utils": "^0.1.3",
-    "@backstage/test-utils": "^0.1.2",
+    "@backstage/cli": "^0.6.6",
+    "@backstage/dev-utils": "^0.1.13",
+    "@backstage/test-utils": "^0.1.9",
     "@spotify/prettier-config": "^8.0.0",
     "@testing-library/jest-dom": "^5.10.1",
     "@testing-library/react": "^10.4.1",
